src/main/java/com/jdom/util/io/IOUtil.java
/** 
 *  Copyright (C) 2012  Just Do One More
 *  
 *  This program is free software: you can redistribute it and/or modify
 *  it under the terms of the GNU General Public License as published by
 *  the Free Software Foundation, either version 3 of the License, or
 *  (at your option) any later version.
 *
 *  This program is distributed in the hope that it will be useful,
 *  but WITHOUT ANY WARRANTY; without even the implied warranty of
 *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *  GNU General Public License for more details.
 *
 *  You should have received a copy of the GNU General Public License
 *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */package com.jdom.util.io;

import java.io.Closeable;

public class IOUtil {

    public static final String NEW_LINE = System.getProperty("line.separator");

    private IOUtil() {

    }

    /**
     * Closes a closeable object. It first checks to see that the object is not null, and if it
     * isn't, it closes it.
     * 
     * @param closeable
     *            The closeable object
     */
    public static void close(Closeable closeable) {
        if (closeable != null) {
            try {
                closeable.close();
            } catch (Exception e) {
                // Eat it
            }
        }
    }
}
